+++ awswrangler/common.py
@@ -5,7 +5,7 @@ def get_session(
     session_primitives=None, key=None, secret=None, profile=None, region=None
 ):
     """
-    Return a configured boto3 Session object
+    Return a configured Boto3 Session object
     """
     if session_primitives:
         key = session_primitives.key if session_primitives.key else key
@@ -28,3 +28,19 @@ class SessionPrimitives:
         self.secret = secret
         self.profile = profile
         self.region = region
+
+
+def calculate_bounders(num_items, num_groups):
+    num_groups = num_items if num_items < num_groups else num_groups
+    size = int(num_items / num_groups)
+    rest = num_items % num_groups
+    bounders = []
+    end = -1
+    for _ in range(num_groups):
+        start = end + 1
+        end += size
+        if rest:
+            end += 1
+            rest -= 1
+        bounders.append((start, end))
+    return bounders
